Fixing 1b30987 - delete member w/ last member role
member will not be saved when all roles are marked for destruction. Therefore
the last member_role will never be removed and consequently the member neither.
Therefore we need to change the tatic.

Explicitly destroying membership in controllers whenever no roles are assigned.

This code needs a serious refactoring. There are three ways to create members
and member_roles and all of them are different. This is destined to cause bugs
and headaches.